requirements.txt
absl-py==0.7.1
astor==0.8.0
bleach==1.5.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.2.3
Flask==1.0.3
Flask-Cors==3.0.8
flask-heroku==0.1.9
Flask-SQLAlchemy==2.4.0
gast==0.2.2
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
html5lib==0.9999999
idna==2.8
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
Pillow==6.0.0
protobuf==3.8.0
psycopg2==2.8.3
pycairo==1.18.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.5
sqlparse==0.3.0
tensorboard==1.8.0
tensorflow==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
